Updated bitwidth

+++ tests/onnx_matrix_mul/test_onnx_matrix_mul_int8.py
@@ -16,9 +16,9 @@ import onnx_matrix_mul
 
 a_shape = (15, 15)
 b_shape = (15, 15)
-a_dtype = ng.int16
-b_dtype = ng.int16
-c_dtype = ng.int16
+a_dtype = ng.int8
+b_dtype = ng.int8
+c_dtype = ng.int8
 par = 1
 axi_datawidth = 32
